added files for password manager
Day 5 project .. Upgrade on the password manager I made back on Day 2. I rebuilt it to be way more organized and easier to work with, header files, classes, and a cleaner layout. The program’s split into separate files for the menu, password creation, and password viewing, so it’s not just one giant block of code anymore. should be much easier to work with & update